setup: add entrypoint for django manage commands
Since the migration to the django framework we need to initialize and/or
migrate the database via the manage.py file before running the
aap-azure-billing command.

This can be done via:

$ python3 manage.py migrate

Because we don't want to pass the full path of the manage.py file (like
/usr/lib/python3.8/site-packages/azure_billing/manage.py from the rpm
package) then we can create a new entrypoint for a better usage.

$ aap-azure-billing-manage migrate
Operations to perform:
  Apply all migrations: billing, main
Running migrations:
  No migrations to apply.

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>